refactoring the twich authentication
